// src/interfaces/actionTypes/personTypes.ts
export const ACTIONS = {
  GET_ALL: "person/get_all",
  GET_COMPANY_PERSONS: "person/get_company_persons",
  GET_PERSONS_BY_TYPE: "person/get_persons_by_type",
  GET: "person/get",
  GET_PERSON_TO_ASSIGN: "person/get_persons_to_assign",
  SET_PERSON: "set_person/get",
  SET_PERSON_ASSIGN_PAGINATION: "person/set_person_assign_pagination",
  GET_FAMILY_BY_PERSON: "person/get_family_by_person",
  GET_PARTNER_STATISTICS: "person/get_parnter_statistics",
  GET_FAMILY_STATISTICS: "person/get_family_statistics",
  GET_GUEST_STATISTICS: "person/get_guest_statistics",
  GET_PERSONS_STATISTICS: "person/get_persons_statistics",
  GET_PERSONS_EXCEPTION_STATISTICS: "person/get_persons_exception_statistics",
  GET_PERSONS_BIRTHDAY_STATISTICS: "person/get_persons_birthday_statistics",
  GET_PERSONS_BIRTHDAY: "person/get_persons_birthday",
  GET_PERSONS_REPORT: "person/get_persons_report",
  SET_LOADING: "person/set_loading",
  SET_COMPANY_PERSONS_LOADING: "person/set_company_persons_loading",
  SET_PERSONS_BY_TYPE_LOADING: "person/set_persons_by_type_loading",
  SET_SECOND_LOADING: "person/set_second_loading",
  SET_RELATION_LOADING: "person/set_relation_loading",
  SET_FAMILY_LOADING: "person/set_family_loading",
  SET_ASSIGN_LOADING: "person/set_assign_loading",
  SET_REPORT_BY_PARTNER_LOADING: "person/set_report_by_partner_loading",
  SET_FAMILIES_PARTNER_CARD_LOADING: "person/set_families_partner_card_loading",
  SET_PERSON_LOCKERS_BY_LOCATION_LOADING: "person/set_person_lockers_by_location_loading",
  SET_PARTNER_STATISTICS_LOADING: "person/set_parnter_statistics_loading",
  SET_FAMILY_STATISTICS_LOADING: "person/set_family_statistics_loading",
  SET_GUEST_STATISTICS_LOADING: "person/set_guest_statistics_loading",
  SET_PERSONS_STATISTICS_LOADING: "person/set_persons_statistics_loading",
  SET_PERSONS_EXCEPTION_STATISTICS_LOADING: "person/set_persons_exception_statistics_loading",
  SET_PERSONS_BIRTHDAY_STATISTICS_LOADING: "person/set_persons_birthday_statistics_loading",
  SET_PERSONS_BIRTHDAY_LOADING: "person/set_persons_birthday_loading",
  SET_PERSON_LOCKERS: "person/set_person_lockers",
  SET_PERSON_LOCKERS_LOADING: "person/set_person_lockers_loading",
  GET_PARTNERS_TO_ASSIGN: "person/get_partners_to_assign",
  GET_TITULAR_TO_ASSIGN: "person/get_titular_to_assign",
  GET_FAMILIES_PARTNER_BY_CARD: "person/get_families_partner_card",
  GET_GUEST_BY_PARTNER: "person/get_guest_by_partner",
  SET_PARTNERS_LOADING: "person/set_partners_loading",
  SET_TITULAR_LOADING: "person/set_titular_loading",
  SET_FAMILY_BY_PERSON_LOADING: "person/set_family_by_person_loading",
  SET_GUEST_BY_PARTNER_LOADING: "person/set_guest_by_partner_loading",
  SET_SELECTED_FAMILIES_BY_PARTNER: "person/set_selected_families_by_partner",
  GET_PERSON_LOCKERS_BY_LOCATION: "person/get_person_lockers_by_location",
  CLEAR: "person/clear",
  CLEAR_PERSONS: "person/clear_persons",
  CLEAR_PERSON_LOCKERS_BY_LOCATION: "person/clear_persons_lockers_by_location",
  SET_PAGINATION: 'person/set_pagination',
};